update:代码封装和抽离
